auth/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
PyJWT==2.8.0
boto3==1.34.82
botocore==1.34.82